Make new functions public and cleanup nodes
